# Conflicts:
#	pom.xml
#	src/main/java/com/ecommerce/g58/controller/ProductController.java
#	src/main/java/com/ecommerce/g58/security/SecurityConfig.java
#	src/main/java/com/ecommerce/g58/service/OrderService.java

+++ src/main/java/com/ecommerce/g58/dto/OrdersDTO.java
@@ -12,6 +12,6 @@ public class OrdersDTO {
     private int orderId;
     private LocalDateTime orderDate;
     private String status;
-    private BigDecimal totalQuantity;
-    private BigDecimal totalPrice;
+    private Integer totalQuantity;
+    private Long totalPrice;
 }
